requirements.txt
chromadb==0.5.5
openai==1.54.4
semantic-kernel==1.12.0
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
pypdf==5.1.0
python-docx==1.1.2
python-pptx==1.0.2
pandas==2.2.3
openpyxl==3.1.5
pillow==11.0.0
pytesseract==0.3.13
beautifulsoup4==4.12.3
lxml==5.3.0
streamlit==1.45.1
